pyproject.toml
[project]
name = "sabre"
version = "0.1.0"
description = "SABRE - Self-Amplifying Backtracking Recursive Execution"
authors = [
    {name = "9600 devs", email = "contact@9600.dev"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core server dependencies
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.30.1",
    "starlette>=0.47.2",
    "python-multipart>=0.0.20",

    # OpenAI Response Executor (needed by: llm_call, llm_bind, coerce, llm_list_bind helpers)
    "openai>=1.98.0",
    "httpx>=0.28.1",
    "tiktoken>=0.9.0",

    # Data handling
    "pydantic>=2.7.4",
    "pyyaml>=6.0.1",
    "orjson>=3.10.4",
    "dill>=0.3.8",
    "pandas>=2.0.0",  # Needed by: pandas_bind helper

    # Client UI (pane)
    "prompt-toolkit>=3.0.47",
    "pygments>=2.18.0",
    "websockets>=12.0",

    # Search
    "ddgs>=9.0.0",

    # Visualization
    "matplotlib>=3.7.0",

    # HTTP client
    "aiohttp>=3.11.13",
    "aiofiles>=23.2.1",
    "requests>=2.32.3",
    "markdownify>=0.11.6",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "PyPDF2>=3.0.0",

    # Browser automation (for JavaScript-heavy sites)
    "playwright>=1.40.0",

    # Utilities
    "click>=8.1.8",
    "colorama>=0.4.6",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2024.1",
    "certifi>=2024.6.2",

    # Common dependencies
    "typing-extensions>=4.12.2",
    "pydantic-core>=2.18.0",
]

[project.urls]
Homepage = "https://github.com/9600dev/llmvm"
Repository = "https://github.com/9600dev/llmvm"

[project.scripts]
sabre = "sabre.cli:main"
sabre-server = "sabre.server.__main__:main"
sabre-client = "sabre.client.__main__:main"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sabre"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.ruff]
select = [
  "F",
]
line-length = 120
target-version = "py312"
